(DO NOT MERGE) Introduce Koboldcpp support
-Presently koboldcpp does not use the openai API format and requires custom coding.  This PR preserves the code that will be necessary to implement koboldcpp (which is used by Herika) if koboldcpp does not provide an openai API endpoint.  Currently the dev of koboldcpp indicates it will support an openai API endpoint on the next release so this PR should not be merged yet.

The PR does the following:

(1) Adds a requirement of the requests module to requirements.txt.  This is needed for the chat_response file for kobold.cpp to create summaries.

(2) Adds an example of adding the default streaming koboldcpp endpoint to the config.ini.

(3) Adds changes to the chat_response.py file necessary to divert summary requests to kobold when user uses kobold.  The fact the user is using kobold is passed through the game manager's llm variable; game manager checks the content of the alternative_openai_api_base variable and if it fits kobold's unique pattern, feeds the endpoint through to character manager and chat_response to be used for the summary.  Chat response uses the non-streaming kobold endpoint since mantella expects a single response with the summary.

(4) adds changes to the output manager (which now needs to import all of aiohttp instead of just ClientSession) to support streaming SSE responses from koboldcpp's api/v1/extra/generate/stream endpoint which follows its own unique SSE query and response format.  It also parses the messages array used by mantella to take out the openai formating and produce a single prompt string, adding ###Instruction and ###Response keywords usually used by a lot of local LLMs.