Misc fixes : NPC lookup and audio files copy
Remove the audio files copy for Fallout 4 in fallout4.py. The wav files are not used anymore since FO4 caches audio in a way that prevent wav file substitutions while the game is running. The wav files also cause crashes when trying to load new NPCs that are absent from the voide model list.
Add an additional try block in prepare_sentence_for_game() for Fallout 4 that attempts creates a missing subdirectory for the voice file in the Mantella.esp directory if it's absent for some reason.
Fix the FEMALE_VOICE_MODELS & MALE_VOICE_MODELS lookup functions so that they correctly append "Race" at the end of the search string for both Fallout4.py & Skyrim.py
Fix the if condition check for actor_race in prepare_sentence_for_game() for Skyrim and Fallout to correctly compared to an integer instead of a string